+++ src/main/java/kuromi/command/AddCommand.java
@@ -1,7 +1,7 @@
 package kuromi.command;
 
-import kuromi.Storage;
-import kuromi.Ui;
+import kuromi.storage.Storage;
+import kuromi.view.Ui;
 import kuromi.task.Task;
 import kuromi.task.TaskList;
 
@@ -13,7 +13,7 @@ public class AddCommand extends Command {
     private Task task;
 
     /**
-     * kuromi.MainWindow.kuromi.KuromiException.Main constructor
+     * kuromi.gui.component.MainWindow.kuromi.exceptions.KuromiException.Main constructor
      *
      * @param task Task to be added.
      */
