+++ src/main/java/kuromi/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package kuromi.command;
 
-import kuromi.Storage;
-import kuromi.Ui;
+import kuromi.storage.Storage;
+import kuromi.view.Ui;
 import kuromi.task.TaskList;
 
 /**
@@ -11,7 +11,7 @@ public class ExitCommand extends Command {
     private CommandType command;
 
     /**
-     * kuromi.MainWindow.kuromi.KuromiException.Main constructor.
+     * kuromi.gui.component.MainWindow.kuromi.exceptions.KuromiException.Main constructor.
      */
     public ExitCommand() {
         this.command = CommandType.bye;
